[refactor][client] Merge coordinator and store client into dingodb_client.
